refactor: narrow staging_transactions.signature_type to UInt8
Aligns staging with the final transactions table. Known signature types
(1=RSA, 2=ed25519, 3=Ethereum, etc.) all fit in UInt8, and any future
out-of-range value will now fail loudly at parquet ingest instead of
silently truncating during staging→final migration.

+++ src/database/clickhouse/schema.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS staging_transactions (
   owner BLOB,
   signature_offset UInt64,
   signature_size UInt64,
-  signature_type UInt32,
+  signature_type UInt8,
   root_transaction_id BLOB,
   root_parent_offset UInt64,
   inserted_at DateTime DEFAULT now(),
